app completed with deletion/updates

+++ randomIdeasApp/routes/ideas.js
@@ -47,17 +47,28 @@ router.post('/', async (req, res) => {
 // update an idea
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+    // match usernames
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true },
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true },
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+    // usernames do not match 
+    res
+      .status(403)
+      .json({
+        success: false,
+        error: 'You are not authorized to delete this resource.',
+      });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
@@ -67,8 +78,19 @@ router.put('/:id', async (req, res) => {
 // delete an idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+    // match usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+    // usernames do not match 
+    res
+      .status(403)
+      .json({
+        success: false,
+        error: 'You are not authorized to delete this resource.',
+      });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
